[RED] add and improve existing tests for pause modal
